# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Work Time Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /todo-lists:
    get:
      operationId: listTodoLists
      tags: [todos]
      summary: List all to-do lists
      responses:
        "200":
          description: List of todo lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoList"
    post:
      operationId: createTodoList
      tags: [todos]
      summary: Create a to-do list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoListInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoList"
  /todo-lists/{id}:
    patch:
      operationId: updateTodoList
      tags: [todos]
      summary: Update a to-do list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoListUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoList"
        "404":
          description: Not found
    delete:
      operationId: deleteTodoList
      tags: [todos]
      summary: Delete a to-do list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /todo-lists/{id}/clear-completed:
    post:
      operationId: clearCompletedTodoTasks
      tags: [todos]
      summary: Soft-clear (hide) all currently-completed tasks in a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                type: object
                required: [clearedCount]
                properties:
                  clearedCount:
                    type: integer
  /todo-tasks:
    get:
      operationId: listTodoTasks
      tags: [todos]
      summary: List tasks, optionally filtered by list
      parameters:
        - name: listId
          in: query
          schema:
            type: integer
          description: "Filter by list"
        - name: projectId
          in: query
          schema:
            type: integer
          description: "Filter by project"
        - name: includeCleared
          in: query
          schema:
            type: boolean
          description: "Include soft-cleared tasks (default false)"
        - name: includeFuture
          in: query
          schema:
            type: boolean
          description: "Include tasks scheduled for a future date (default false)"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoTask"
    post:
      operationId: createTodoTask
      tags: [todos]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"
  /todo-tasks/calendar-summary:
    get:
      operationId: getTodoCalendarSummary
      tags: [todos]
      summary: Per-day completion rates for all lists in a date range
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
          description: "YYYY-MM-DD"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
          description: "YYYY-MM-DD"
      responses:
        "200":
          description: Array of daily list summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoCalendarSummaryItem"
  /todo-tasks/day-detail:
    get:
      operationId: getTodoDayDetail
      tags: [todos]
      summary: Reconstructed task list + completion state for a list on a specific date
      parameters:
        - name: listId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
          description: "YYYY-MM-DD"
      responses:
        "200":
          description: Tasks as they stood on that date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoDayDetailTask"
  /todo-tasks/day-detail/toggle:
    post:
      operationId: toggleTodoDayDetailTask
      tags: [todos]
      summary: Toggle a task's completion for a specific (typically past) date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoDayDetailToggleInput"
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoDayDetailTask"
  /todo-tasks/{id}:
    patch:
      operationId: updateTodoTask
      tags: [todos]
      summary: Update a task's text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"
        "404":
          description: Not found
    delete:
      operationId: deleteTodoTask
      tags: [todos]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /todo-tasks/{id}/complete:
    post:
      operationId: completeTodoTask
      tags: [todos]
      summary: Mark a task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"
        "404":
          description: Not found
  /todo-tasks/{id}/uncomplete:
    post:
      operationId: uncompleteTodoTask
      tags: [todos]
      summary: Mark a task as incomplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task uncompleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"
        "404":
          description: Not found
...
components:
  schemas:
...
    TodoList:
      type: object
      required: [id, name, color, letter, resetDaily, autoClearCompleted, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        letter:
          type: string
          description: "Single uppercase character shown in badges"
        resetDaily:
          type: boolean
        autoClearCompleted:
          type: boolean
          description: "If true, completed tasks are hidden (soft-cleared) once the day ends"
        sortOrder:
          type: integer
        createdAt:
          type: string
    TodoListInput:
      type: object
      required: [name, color, letter]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        letter:
          type: string
          minLength: 1
          maxLength: 1
        resetDaily:
          type: boolean
        autoClearCompleted:
          type: boolean
        sortOrder:
          type: integer
    TodoListUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        letter:
          type: string
          minLength: 1
          maxLength: 1
        resetDaily:
          type: boolean
        autoClearCompleted:
          type: boolean
        sortOrder:
          type: integer
    TodoTask:
      type: object
      required: [id, listId, text, sortOrder, createdAt]
      properties:
        id:
          type: integer
        listId:
          type: integer
        projectId:
          type: ["integer", "null"]
        subprojectId:
          type: ["integer", "null"]
        text:
          type: string
        completedAt:
          type: ["string", "null"]
        completedDate:
          type: ["string", "null"]
          description: "YYYY-MM-DD of completion; used for daily-reset logic"
        clearedAt:
          type: ["string", "null"]
          description: "Soft-hide timestamp; cleared tasks no longer show on the live list"
        scheduledDate:
          type: ["string", "null"]
          description: "YYYY-MM-DD; if set and in the future, task is 'prepared' and hidden until then"
        reminderTime:
          type: ["string", "null"]
          description: "HH:mm; opt-in reminder time of day"
        sortOrder:
          type: integer
        createdAt:
          type: string
    TodoTaskInput:
      type: object
      required: [listId, text]
      properties:
        listId:
          type: integer
        text:
          type: string
          minLength: 1
        projectId:
          type: integer
        subprojectId:
          type: integer
        scheduledDate:
          type: string
          description: "YYYY-MM-DD"
        reminderTime:
          type: string
          description: "HH:mm"
    TodoTaskUpdate:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
        subprojectId:
          type: ["integer", "null"]
        scheduledDate:
          type: ["string", "null"]
        reminderTime:
          type: ["string", "null"]
        listId:
          type: integer
          description: "Move the task to a different list"
        sortOrder:
          type: integer
    TodoCalendarSummaryItem:
      type: object
      required: [date, listId, listName, listColor, letter, totalTasks, completedTasks, preparedTasks, percentage]
      properties:
        date:
          type: string
          description: "YYYY-MM-DD"
        listId:
          type: integer
        listName:
          type: string
        listColor:
          type: string
        letter:
          type: string
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        preparedTasks:
          type: integer
          description: "Count of tasks scheduled (prepared in advance) for this date"
        percentage:
          type: integer
          description: "0-100"
    TodoDayDetailTask:
      type: object
      required: [taskId, text, completed]
      properties:
        taskId:
          type: integer
        text:
          type: string
        completed:
          type: boolean
    TodoDayDetailToggleInput:
      type: object
      required: [taskId, date, completed]
      properties:
        taskId:
          type: integer
        date:
          type: string
          description: "YYYY-MM-DD"
        completed:
          type: boolean
...