# lib/api-zod/src/generated/types/todoTaskUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TodoTaskUpdate {
  /** @minLength 1 */
  text?: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  subprojectId?: number | null;
  /** @nullable */
  scheduledDate?: string | null;
  /** @nullable */
  reminderTime?: string | null;
  /** Move the task to a different list */
  listId?: number;
  sortOrder?: number;
}